--- a/chia/cmds/show_funcs.py
+++ b/chia/cmds/show_funcs.py
@@ -5,6 +5,7 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 
 from chia.rpc.full_node_rpc_client import FullNodeRpcClient
+from chia.types.blockchain_format.sized_bytes import bytes32
 
 
 async def print_blockchain_state(node_client: FullNodeRpcClient, config: Dict[str, Any]) -> bool:
@@ -36,7 +37,7 @@ async def print_blockchain_state(node_client: FullNodeRpcClient, config: Dict[st
 
     if synced:
         print("Current Blockchain Status: Full Node Synced")
-        print("\nPeak: Hash:", peak.header_hash if peak is not None else "")
+        print("\nPeak: Hash:", bytes32(peak.header_hash) if peak is not None else "")
     elif peak is not None and sync_mode:
         sync_max_block = blockchain_state["sync"]["sync_tip_height"]
         sync_current_block = blockchain_state["sync"]["sync_progress_height"]
@@ -44,7 +45,7 @@ async def print_blockchain_state(node_client: FullNodeRpcClient, config: Dict[st
             f"Current Blockchain Status: Syncing {sync_current_block}/{sync_max_block} "
             f"({sync_max_block - sync_current_block} behind)."
         )
-        print("Peak: Hash:", peak.header_hash if peak is not None else "")
+        print("Peak: Hash:", bytes32(peak.header_hash) if peak is not None else "")
     elif peak is not None:
         print(f"Current Blockchain Status: Not Synced. Peak height: {peak.height}")
     else:
@@ -55,7 +56,7 @@ async def print_blockchain_state(node_client: FullNodeRpcClient, config: Dict[st
         if peak.is_transaction_block:
             peak_time = peak.timestamp
         else:
-            peak_hash = peak.header_hash
+            peak_hash = bytes32(peak.header_hash)
             curr = await node_client.get_block_record(peak_hash)
             while curr is not None and not curr.is_transaction_block:
                 curr = await node_client.get_block_record(curr.prev_hash)
@@ -84,7 +85,7 @@ async def print_blockchain_state(node_client: FullNodeRpcClient, config: Dict[st
             curr = await node_client.get_block_record(curr.prev_hash)
 
         for b in added_blocks:
-            print(f"{b.height:>9} | {b.header_hash}")
+            print(f"{b.height:>9} | {bytes32(b.header_hash)}")
     else:
         print("Blockchain has no blocks yet")
     return False