This PR can be reviewed one commit at a time, for bite-sized changes.

# Purpose:

This patch transitions `FullBlock`, `UnfinishedBlock` and `BlockRecord` over to use the rust counterparts.
The rust types are defined [here](https://github.com/Chia-Network/chia_rs/blob/main/chia-protocol/src/fullblock.rs), [here](https://github.com/Chia-Network/chia_rs/blob/main/chia-protocol/src/unfinished_block.rs) and [here](https://github.com/Chia-Network/chia_rs/blob/main/chia-protocol/src/block_record.rs).

The most immediate consequence of this is that serialization and deserialization of these types is a lot efficient. This addresses the following issues:

* When validating blocks, we serialize them in order to pass them to the worker process (and deserialize them in the worker process).
* When requesting weight proofs, we deserialize many `BlockRecord` objects from the database

The long term goal is to move more of the block validation into rust, and worker *threads* instead of processes.

As usual with the transitions to rust types, they are not `dataclass`es, so calls to `replace()` has to be changed to a member function rather than the dataclasses free function.

# Current Behavior:

parsing `FullBlock`, `UnfinishedBlock` and `BlockRecord` is slow.

# New Behavior:

parsing `FullBlock`, `UnfinishedBlock` and `BlockRecord` is faster. (see benchmark below).

# Benchmarks

## `benchmarks/block_store.py`

before:

```
all tests completed in 367.2139s
database size: 93.503 MB
```

after:

```
all tests completed in 353.3438s
database size: 93.503 MB
```

| test | before | after | after / before |
| --- | --: | --: | --: |
|add_full_block                    |  232.27s| 229.05s | 98.60% |
|get_block_info                    |  5.30s  | 4.35s   | 82.20% |
|get_generator                     |  4.80s  | 4.05s   | 84.46% |
|get_full_block                    |  16.65s | 16.85s  | 101.2% |
|get_full_block_bytes              |  5.53s  | 5.03s   | 90.97% |
|get_generators_at                 |  1.98s  | 2.11s   | 106.25% |
|get_full_blocks_at                |  17.13s | 16.67s  | 97.26% |
|**get_block_records_by_hash**         |  **8.01s**  | **5.12s**   | **63.90%** |
|get_block_bytes_by_hash           |  6.56s  | 5.62s   | 85.60% |
|get_blocks_by_hash                |  17.85s | 16.66s  | 93.40%|
|**get_block_record**                  |  **7.69s**  | **5.14s**   | **66.80%**|
|**get_block_records_in_range**        |  **0.4031s**  | **0.1301s**   |**32.30%** |
|**get_block_records_close_to_peak**   |  **0.0044s**  | **0.0022s**   | **50.00%** |
|is_fully_compactified             |  11.20s | 9.85s   | 88.00% |
|get_random_not_compactified       |  31.83s | 32.72s  |102.80%|


## `benchmarks/blockchains.py`

This benchmark loads from disk all the test chains we use in out tests (the fixtures).

| before | after | after / before |
| --- | --- | --- |
| 34.07s | 10.95s | 32.14% |

## main

![load-test-chain](https://github.com/Chia-Network/chia-blockchain/assets/661450/f6dae068-28d9-47d5-ab36-6a57a17cb2f9)

## This patch

![load-test-chain](https://github.com/Chia-Network/chia-blockchain/assets/661450/7f368062-ac69-43ee-9e79-fe3701480cde)


## profile of requesting weight proofs

![chia-hotspot-22817](https://github.com/Chia-Network/chia-blockchain/assets/661450/38a97210-c699-4c7e-aee1-e66b51e2a81d)

![chia-hotspot-13555-13558](https://github.com/Chia-Network/chia-blockchain/assets/661450/a214fa3f-a247-4a55-b446-2de28eab9f7f)